refactor: split Checklist into sub-components and hooks (#4088)
