+++ editor.planx.uk/src/@planx/components/Checklist/Public/tests/mockOptions.ts
@@ -1,7 +1,8 @@
 import { Option } from "@planx/components/shared";
 
 import { Group } from "../../model";
-import { ChecklistLayout } from "../Public";
+import { ChecklistLayout } from "../components/VisibleChecklist";
+
 
 export const options: {
   [key in ChecklistLayout]?: Array<Option>;
@@ -65,7 +66,7 @@ export const options: {
 };
 
 export const optionsWithExclusiveOption: Array<Option> = [
-  ...options[ChecklistLayout.Basic] || [],
+  ...(options[ChecklistLayout.Basic] || []),
   { id: "tent_id", data: { text: "Tent", exclusive: true } },
 ];
